+++ Tests/Translation/Extractor/File/TranslationContainerExtractorTest.php
@@ -23,6 +23,8 @@ use JMS\TranslationBundle\Model\FileSource;
 use JMS\TranslationBundle\Model\Message;
 use JMS\TranslationBundle\Translation\Extractor\File\TranslationContainerExtractor;
 use JMS\TranslationBundle\Model\MessageCatalogue;
+use PhpParser\Lexer;
+use PhpParser\ParserFactory;
 
 class TranslationContainerExtractorTest extends \PHPUnit_Framework_TestCase
 {
@@ -53,8 +55,14 @@ class TranslationContainerExtractorTest extends \PHPUnit_Framework_TestCase
             $extractor = new TranslationContainerExtractor();
         }
 
-        $lexer = new \PHPParser_Lexer();
-        $parser = new \PHPParser_Parser($lexer);
+        $lexer = new Lexer();
+        if(class_exists('PhpParser\ParserFactory')) {
+            $factory = new ParserFactory();
+            $parser = $factory->create(ParserFactory::PREFER_PHP7,$lexer);
+        } else {
+            $parser = new \PHPParser_Parser($lexer);
+        }
+
         $ast = $parser->parse(file_get_contents($file));
 
         $catalogue = new MessageCatalogue();