+++ Translation/Extractor/File/DefaultPhpFileExtractor.php
@@ -28,6 +28,11 @@ use JMS\TranslationBundle\Annotation\Ignore;
 use JMS\TranslationBundle\Translation\Extractor\FileVisitorInterface;
 use JMS\TranslationBundle\Model\MessageCatalogue;
 use JMS\TranslationBundle\Logger\LoggerAwareInterface;
+use PhpParser\Comment\Doc;
+use PhpParser\Node;
+use PhpParser\NodeTraverser;
+use PhpParser\NodeVisitor;
+use PhpParser\Node\Scalar\String_;
 use Psr\Log\LoggerInterface;
 
 /**
@@ -37,7 +42,7 @@ use Psr\Log\LoggerInterface;
  *
  * @author Johannes M. Schmitt <schmittjoh@gmail.com>
  */
-class DefaultPhpFileExtractor implements LoggerAwareInterface, FileVisitorInterface, \PHPParser_NodeVisitor
+class DefaultPhpFileExtractor implements LoggerAwareInterface, FileVisitorInterface, NodeVisitor
 {
     private $traverser;
     private $catalogue;
@@ -53,7 +58,7 @@ class DefaultPhpFileExtractor implements LoggerAwareInterface, FileVisitorInterf
     public function __construct(DocParser $docParser)
     {
         $this->docParser = $docParser;
-        $this->traverser = new \PHPParser_NodeTraverser();
+        $this->traverser = new NodeTraverser();
         $this->traverser->addVisitor($this);
     }
 
@@ -65,10 +70,10 @@ class DefaultPhpFileExtractor implements LoggerAwareInterface, FileVisitorInterf
         $this->logger = $logger;
     }
 
-    public function enterNode(\PHPParser_Node $node)
+    public function enterNode(Node $node)
     {
 
-        if (!$node instanceof \PHPParser_Node_Expr_MethodCall
+        if (!$node instanceof Node\Expr\MethodCall
             || !is_string($node->name)
             || ('trans' !== strtolower($node->name) && 'transchoice' !== strtolower($node->name))) {
 
@@ -79,7 +84,7 @@ class DefaultPhpFileExtractor implements LoggerAwareInterface, FileVisitorInterf
         $ignore = false;
         $desc = $meaning = null;
         if (null !== $docComment = $this->getDocCommentForNode($node)) {
-            if ($docComment instanceof \PhpParser\Comment\Doc) {
+            if ($docComment instanceof Doc) {
                 $docComment = $docComment->getText();
             }
             foreach ($this->docParser->parse($docComment, 'file '.$this->file.' near line '.$node->getLine()) as $annot) {
@@ -93,7 +98,7 @@ class DefaultPhpFileExtractor implements LoggerAwareInterface, FileVisitorInterf
             }
         }
 
-        if (!$node->args[0]->value instanceof \PHPParser_Node_Scalar_String) {
+        if (!$node->args[0]->value instanceof String_) {
             if ($ignore) {
                 return;
             }
@@ -112,7 +117,7 @@ class DefaultPhpFileExtractor implements LoggerAwareInterface, FileVisitorInterf
 
         $index = 'trans' === strtolower($node->name) ? 2 : 3;
         if (isset($node->args[$index])) {
-            if (!$node->args[$index]->value instanceof \PHPParser_Node_Scalar_String) {
+            if (!$node->args[$index]->value instanceof String_) {
                 if ($ignore) {
                     return;
                 }
@@ -148,12 +153,12 @@ class DefaultPhpFileExtractor implements LoggerAwareInterface, FileVisitorInterf
     }
 
     public function beforeTraverse(array $nodes) { }
-    public function leaveNode(\PHPParser_Node $node) { }
+    public function leaveNode(Node $node) { }
     public function afterTraverse(array $nodes) { }
     public function visitFile(\SplFileInfo $file, MessageCatalogue $catalogue) { }
     public function visitTwigFile(\SplFileInfo $file, MessageCatalogue $catalogue, \Twig_Node $ast) { }
 
-    private function getDocCommentForNode(\PHPParser_Node $node)
+    private function getDocCommentForNode(Node $node)
     {
         // check if there is a doc comment for the ID argument
         // ->trans(/** @Desc("FOO") */ 'my.id')